+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## 2.3.0 (upcoming)
+## 2.3.0 (2024-07-01)
 
 ![Bento Grid of version 2.3.0](https://aplos.gxbs.me/images/v2-3-0.png)